Move task-related files into bob.task pkg
