+++ src/main/java/bob/command/DeadlineCommand.java
@@ -2,9 +2,9 @@ package bob.command;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
-import bob.Deadline;
-import bob.TaskList;
 import bob.Ui;
+import bob.task.Deadline;
+import bob.task.TaskList;
 
 public class DeadlineCommand extends Command {
