+++ src/main/java/bob/task/Event.java
@@ -1,5 +1,7 @@
-package bob;
+package bob.task;
 import java.time.LocalDateTime;
+
+import bob.Helper;
 
 public class Event extends Task {
     private LocalDateTime from;
