+++ src/main/java/bob/task/TaskList.java
@@ -1,6 +1,8 @@
-package bob;
+package bob.task;
 import java.io.IOException;
 import java.util.List;
+
+import bob.Storage;
 
 public class TaskList {
     private static List<Task> tasks;
