+++ src/main/java/bob/task/Todo.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.task;
 public class Todo extends Task {
     public Todo (String description) {
         super(description);
